requirements.txt
Pillow>=8.3.1
wand>=0.5.9
requests>=2.23.0
numpy>=1.23.1
scipy~=1.9.1
setuptools~=58.1.0
sphinx-rtd-theme
sphinx-autodoc-typehints